Filter Samples based on Library Size

This function is a part of the data wrangling functionality of tcgaCleaneR. It allows user to handle the bias in SummarizedExperiment S4 class Cancer Dataset (e.g. TCGA dataset) due to library size by filtering out the samples with sample size greater than the threshold. Using plotLibSize, user can determine the threshold.

filterSamplesByLibSize(data, ls_cutoff)

Arguments

data

SummarizedExperiment S4 class Dataset. E.g. TCGA Dataset.

ls_cutoff

numeric: library size threshold

Value

S4 data object

Examples

filterSamplesByLibSize(data = brca.data, ls_cutoff = 17.5)
#> class: SummarizedExperiment 
#> dim: 100 1126 
#> metadata(0):
#> assays(3): HTseq_counts HTseq_FPKM HTseq_FPKM.UQ
#> rownames(100): TSPAN6 TNMD ... CROT ABCB4
#> rowData names(9): EnsemblGene_ids Gene_symbol ...
#>   NanostringPanCancer_HK sinscorePanCancer_HK
#> colnames(1126): TCGA-A8-A06Z-01A-11R-A00Z-07
#>   TCGA-A8-A08F-01A-11R-A00Z-07 ... TCGA-LD-A9QF-01A-32R-A41B-07
#>   TCGA-BH-A0H9-11A-22R-A466-07
#> colData names(12): Samples Tissues ... CPE Subtypes