Generate RUV-III Data Object

This function is a part of the data analysis functionality of tcgaCleaneR. It captures both the uses PRPS values from library tcgaCleaneR combined with row counts and run SVD algorithm (runSVD()) from BiocSingular on the combined dataset. The function uses RUV-I algorithm from ruv as a pre-processing step to RUV-III.

runRUVIII(
  ruv.data,
  ruv.rep,
  ncg.set,
  k = NULL,
  eta = NULL,
  svd_k = 50,
  include.intercept = TRUE,
  BPPARAM = SerialParam(),
  BSPARAM = ExactParam(),
  fullalpha = NULL,
  return.info = FALSE,
  inputcheck = TRUE
)

Arguments

ruv.data
S4 data object for RUV-III: A S4 data object with combined data including the row count from original filtered data using assay HTseq_counts, prps data for batch and library size. This data needs to be further converted to log scale and transposed.

ruv.rep
S4 data matrix for RUV-III: A S4 data object that has been generated using replicate.matrix functionality from ruv package. This helps ruv to identify replicate samples.

ncg.set
logical object: Set of Negative Controlled genes.

k
Integer scalar specifying the number of. Default is NULL. Currently Value 1 represents the library size, 2 represents purity and 3 is time variation.

eta
Gene-wise (as opposed to sample-wise) covariates. A matrix with n columns for ruv::RUV1. Default is NULL.

svd_k
Integer scalar specifying the number of singular values to return for BiocSingular::runSVD.Default is 50.

include.intercept
Add an intercept term to eta if it does not include one already for ruv::RUV1. Default is True.

BPPARAM
A BiocParallelParam object specifying how parallelization should be performed. Default is SerialParam()

BSPARAM
A BiocSingularParam object specifying the type of algorithm to run. Default is ExactParam()

fullalpha
To perform RUV-III calculation. Default is NULL.

return.info
logical: Do you want all the information related to RUV-III object. False gives all information whereas True gives only the

inputcheck
logical: Check the inputs to identify if ruv.data contains missing values or infinite values.

Value

Based on the return.info we get either a S4 list will all information related to RUV-III object or just the RUV-III result.

Examples

if (FALSE) {
runRUVIII(ruv.data = ruv.data, ruv.rep = ruv.rep, ncg.set = ncg.set, k=1, BSPARAM = BiocSingular::bsparam(), return.info = TRUE)
}